pyproject.toml
[tool.poetry]
name = "blade-defect-detection"
version = "0.1.0"
description = "Blade defect detection using deep learning"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
torchvision = "^0.15.0"
pytorch-lightning = "^2.0.0"
hydra-core = "^1.3.0"
omegaconf = "^2.3.0"
mlflow = "^2.8.0"
pillow = "^10.0.0"
numpy = "^1.24.0"
fire = "^0.5.0"
dvc = "^3.0.0"
torchmetrics = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
ruff = "^0.1.0"
pre-commit = "^3.0.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

